Adds first pass at `new-order` support.
This commit adds a first pass at new-order support. To do so a few new
types are introduced, as well as a `/new-order` and `/my-order`
endpoints for creating & retrieving orders. No authzs are created yet
and the CSR is parsed for structure but not validated against any
issuance policies.

The in-memory datastore is updated to support adding orders and looking
them up by ID (random strings as the ID for now).

To help iterate on development a simple pebble-client was added that can
be a REPL around JWT singing JSON bodies and POSTing them to Pebble.