Refactor internal types.
This commit reworks the existing types slightly such that the `acme`
package types are used strictly for the protocol messages. A new `core`
package is used for all of the Pebble types that contain parsed values
and fields not echoed to a user directly (e.g. ID fields).

The memorystore is updated to only process core types. The WFE is
updated to create core types out of the acme type read off of the wire
as appropriate. Each internal type has a struct embedding of the acme
type allowing the WFE to output the acme type from a core type as
required (e.g. as the reply to a POST for creating a resource).