Support HTTP-01 challenge validation. (#8)
This commit adds initial support for HTTP-01 challenge validation. No
other challenge types are implemented at this time.

The implementation is largely a pared down version of what Boulder does
presently. For simplicities sake there are a number of important
differences that make Pebble *absolutely* unappropriate for production
uses:
  1) The validation is performed in the same goroutine as the WFE's HTTP
     handler instead of in a separate go routine. It might be worth
     revamping this in the near future.
  2) There is no read limit on the HTTP request body and timeouts are
     not as strict
  3) The validity of both pending and valid authzs is hardcoded at 1h.

As compared to Boulder the Pebble challenge design chooses to use
a separate endpoint (e.g. `/chalZ/9999` for chal 9999) instead of
referring to challenges subindexed from authorizations (e.g.
`/authZ/9999/1` for the first challenge of authz 9999). I believe this
is still within compliance with the overall specification and will
further exercise clients to not expect Boulder's choices to be
universal. The pebble client is updated to allow specifying a URL
to POST in addition to a directory endpoint to allow POSTing these
separate non-directory defined challenge URLs.