Embeds certificate ID in order, not path prefix.
Based on review feedback this commit updates the CA to stash
a certificate ID in the internal order object instead of a path prefix.
The WFE then constructs a certificate URL for the certificate based on
the ID in the Order() handler. This keeps public facing paths as a WFE
responsibility.