Adds locking, reworks concurrency based on review feedback.
This commit updates the core internal types to embed RWMutex's to
protect against concurrent reads/updates. The relevant wfe, db, va, and
ca functions were updated to acquire read/write locks as appropriate.

This commit also reworks parts of the CA and VA to have shorter, easier
to read functions.

The VA is updated to not serialize the validation work as it reads
vaTasks off of a channel. The WaitGroup was removed and results from
validation attempts are processed as they occur.