Replaces `Order.CertID` with pointer to object.
Per review feedback its strange to have two IDs when one is just used to
indicate that an order is complete. Instead the internal Order type
should hold a pointer to the internal Certificate type. The ID and DER
encoding can be read from that as required for delivering the order.