Adds TLS-SNI-02 challenge support.
This commit adds TLS-SNI-02 challenges as a supported challenge type in
Pebble. The WFE is modified to construct a TLS-SNI-02 challenge along
with an HTTP-01 challenge for every authz. A new configuration parameter
is introduced to control what port the VA uses for TLS, allowing Chisel
to eventually stand up a TLS-SNI-02 server that won't need root to bind
443.

The validation implementation is largely identical to Boulder (thanks
calavera!) with modifications for Pebble's VA & some slight
simplifications where possible.