Fixes TLS-SNI-02 and DNS-01 key auth calculations.
The `KeyAuthorization` field on the `task.Challenge` object is for
a KeyAuthorization provided by the client as part of the HTTP-01
challenge POST body. It will be empty when calculating the expected
value for a TLS-SNI-02 or DNS-01 challenge.

Instead the server should construct the expected `KeyAuthorization`
value by calling the challenge's `ExpectedKeyAuthorization` function
with the registration owning the challenge.