Switch to lookup-based validation.
The latest version of the ACME spec requires all requests other than
registration to be authenticated by looking up an account key via the
"kid" JWS header parameter. This change implements that by defining two methods
to get keys from a JWS: one that extracts the key from the "jwk" header field
(for use during registration) and one that extracts it from the "kid" header
field (for use with all other requests). It also enforces that the two header
fields are mutually exclusive.