Allow disabling VA sleeps using PEBBLE_VA_NOSLEEP env var. (#35)
This commit adds support for a `PEBBLE_VA_NOSLEEP` env var. When set to `1` (or "true", "True" to be user friendly) the Pebble VA will not sleep a random amount of time between validation attempts. The README is updated to describe the default sleep behaviour and to add a description of using this env var to test validation at full speed without artificial delays.

This design is informed by the review on my previous (closed) PR: #34