Don't hold order lock while iterating authorizations.
Prior to this commit `VA.maybeIssue` held a `RLock` on the order object
it was checking for completeness while it iterated over each of the
authorizations associated with the order, locking them in turn.

This commit changes this behaviour to instead acquire a read lock on the
order only to read the `[]*Authorization` field of the order before
freeing the lock for other readers. This allows the iteration over the
authorizations to continue independent of the order. The
`CA.CompleteOrder` function will lock it again before modifying if the
order was in fact complete and ready to issue.

This fixes the deadlock observed when submitting a CSR with more than
one name.