Use chisel2.py from Boulder repo for CI.
The `chisel2.py` integration test client was moved out of the Certbot
repo into Boulder. This makes for an interesting dance for CI on Pebble.
We still have to clone the Certbot repo at the acme-v2 branch to get
the right `acme` python module and the certbot deps. We also have to get
the `chisel2.py` python script from the Boulder repo. Rather than clone
the whole repo I've opted to have CI pull the raw asset it requires
directly from master.