Fix JWS auth type enforcement error messages.
Previous to this commit the checks done in `parseJWS` and `lookupJWK`
would return a confusing "Mutually exclusive" error message if the wrong
auth type were used.

This commit rearranges the checks such that the mutually exclusive error
only happens when two auth types are used together. When the *wrong*
auth type is used in isolation a much more sensible error is returned.