Do not expect/validate "Type" in challenge POST.
Previous to this commit the Pebble `validateChallengeUpdate` function
assumed that the POST JWS payload sent to the challenge update endpoint
contained both "type" and "keyAuthorization" fields. If the provided
"type" value didn't match the server's understanding of the challenge
type an error was returned.

The current ACME draft specification does not require this "type" field.
Section 7.5.1 only describes sending a "keyAuthorization". This
commit removes the expectation that a type field is sent, and the
constraint that it be equal to the expected challenge type.