Implement wildcard issuance policy. (#80)
This commit adds a simplified duplication of the Boulder wildcard
issuance policy[0] for Pebble.

This primarily consists of allowing identifiers to take values with
a wildcard prefix ("*.") internally. When an authorization for an
identifier with a wildcard issuance is created only a DNS-01 challenge
is added to the authorization. The WFE and VA strip the "*." prefix from
the identifier as required for presentation to the user (to match ACME
semantics) and for validation.

[0]
- https://github.com/letsencrypt/boulder/commit/1c99f91733ec136f205e6a192f1abe0ae1c3f342