Add config flag to enable/disable strict ACME enforcement.
Prior to this commit Pebble enforced both that no `keyAuthorization` was
sent in Challenge POST bodies and that all JWS POSTs included the
correct `Content-Type`.

The Let's Encrypt staging/prod V2 environments do not support *not*
sending the `keyAuthorization` (pending an update soon to be deployed),
while Pebble would not support sending it. This leaves clients unable to
support both Pebble and Boulder.

As a compromise this commit introduces a "-strict" flag. When enabled,
breaking changes are enforced by Pebble. This gives client developers an
easy way to check for future API breaking changes. For now we default to
off. After the Boulder updates we should instead default to "on".