WFE: Remove unused challenges for non-pending authz.
This commit updates the WFE's `prepAuthorizationForDisplay` function
such that if the authorization being prep'd is not status pending then
the associated challenges are filtered to only those that were used to
make the authz change status. In effect this means that an invalid or
valid authz will only have the challenge that was POSTed included in the
authz output.

This better matches what the ACME draft says about the "challenges"
field in Section 7.1.4 "Authorization Objects".