VA: Allow skipping real challenge validation requests.
An ACME client developer requested a way to use Pebble to test ACME
interactions without having to go through the complexities of standing
up a real challenge response server. This lets the developer focus on
the baseline ACME issuance flow without needing to actually respond to
HTTP-01, DNS-01, or TLS-ALPN-01 challenge requests.

This commit adds a new env var `PEBBLE_VA_ALWAYS_VALID` to accomplish
the above. When enabled the VA will not actually perform validation when
a challenge is POSTed and instead will assume the challenge is valid.