Handle order status as a property of assoc. authz status.
This commit updates the way Pebble handles an Order's status field to
closer match the way we handle it in Boulder. This means the order's
status is considered to be a calculated field derived based on the
order's error/beganProcessing/certificate fields and the
status/expires/error of the order's associated authorizations. Along the
way the new "Ready" status was implemented. This commit also addresses
the problem of the order status not being updated when an authorization
is failed. Now, an order has a problem details field that is set when an
authorization associated with the order has a failed challenge
validation.