WFE: Return malformed problem for revoke cert 404.
Prior to this commit if a revocation request specified a certificate
that Pebble did not have in the memorystore a bare HTTP 404 response was
returned. It would be better to return an ACME problem (like Boulder
does) in this case so this commit updates the WFE to return a malformed
problem with an appropriate message.