WFE: Fix revocation authentication.
Prior to this commit the revocation implementation specified it only
supported authenticating the request with the certificate's key and an
embedded JWK. In practice the revocation logic was still expecting a key
ID and authentication from the ACME account.

This commit fixes the revocation implementation to match what is
described: it reads the JWK from the request JWS and expects it to match
the certificate's public key.

We should additionally support authenticating the request with the ACME
account credentials of the issuer but this is left for a follow-up.