Rework revocation to support both by-account and by-cert-key.
This commit ports over more of the ACMEv2 revocation processing from
Boulder. With this code in place it is possible for an account to revoke
a certificate it previously issued using a keyID authenticated JWS. It
also allows any account to revoke a certificate by posting the
revocation endpoint with a JWS authenticated with an embedded JWK
carrying the certificate's public key.

TODO: Implement revocation of a certificate issued by a different
account when the revoking account has full authorization for all of the
certificate's identifiers.