Distinguish between empty contact list and not specified contact list. (#149)
When updating an account, this PR changes the behavior such that an empty contact list is treated differently than a not specified one (or one which is set to null, which cannot be distinguished easily from a not specified one). This allows clients to actually delete the contacts list.

Fixes #114.