Initial module support w/ vendored deps (#173)
This PR updates Pebble to be a Go 1.11 module. See [the Go wiki page on modules](https://github.com/golang/go/wiki/Modules) for more information.

All of the dependencies are now vendored. CI is updated to explicitly use the vendored dependencies instead of cloning the latest versions at runtime. This should improve the reliability and reproducibility of our builds. This shouldn't affect traditional usage of Pebble in a `$GOPATH` which will continue to work as before.