Use pre-build release versions of pebble and pebble-challtestsrv in docker-compose.yml (#216)
This PR modifies this file to use the latest released version in Docker Hub,
like it is done for Boulder. This allows to use a fully tested and released
version of Pebble, that will be a benefit for the Certbot CI workflow, and for
any user that would like to use the `docker-compose.yml` file in the GIT repo as
a fast and reliable setup.