core: remove Account ID field from JSON responses. (#225)
The `core.Account` type has an `ID` string field that is used to
internally track the ID of the ACME account. It should not be returned
in JSON `Account` resource responses, it's not an RFC 8555 specified
field. Boulder may return a similar field but that's a convenience and
not a protocol feature. The RFC 8555 method of learning the account's ID
is from the `Location` header.