fix dataraces, add -race install, add load-gen.
The `Authz()` method of the WFE was racey. One because it didn't lock
the authorizations and orders it was working with. Second because the
`acme.Authorization` type had pointers to challenge objects that were
being manipulated by the VA. Both of these races are fixed in this
branch.

To prevent future occurrences travis.yml is updated to install Pebble
with the race detector enabled, and to run Pebble such that it will exit
non-zero if a race is detected.

Since Chisel2.py is single threaded the Boulder load-generator is used
for a short duration to drive concurrent request traffic. In practice
before fixing the dataraces I found this would crash Pebble <30s.