CI: publish release assets to GitHub (#240)
Sets up automated deployment of executable assets to Pebble GitHub releases.

Travis CI and AppVeyor are configured to take advantage of their GitHub release
deploy provider. It allows to setup upon a tag build the proper creation of
a GitHub release draft (if it does not exist yet), then create/update the
relevant pebble/challtestsrv executables to this release.

Note that Travis handles the Linux executables, and AppVeyor the Windows ones.
In theory, thanks to the capability of Go to cross-compile, one CI system coul
do all the job. However here, I ensure that the platform-specific assets are not
pushed to the release before they have been fully tested in their native
platform. I think this is desirable and so I setup the workflow accordingly.

Once done, upon a tag push to GitHub (like v2.0.3), the two CI will trigger the
tag build, and deploy the artifacts at the end.

Note again that the release is set as a Draft, and so do not become public
before a manual action is done, like editing the release body to push
a changelog. Note also that the two CIs works independently from each other, and
reuse the existing draft release if exists, without trying to recreate the
release (that would drop the assets pushed by the other CI).