CA: allow configuring EE cert OCSP Responder URL (#242)
Pebble is not designed to be an OCSP responder of the certificates it issues,
since it is not its purpose. However, it is easy to setup a proper OCSP
responder alongside it, thanks to the ability to get the intermediate CA
certificate with `/intermediate` and `/intermediate-key` endpoints.

To make this interaction complete, and allow a client (like Certbot) to get
a proper OCSP response status during tests, the certificates generated by Pebble
need to contain an arbitrary OCSP Responder URL.

This PR allows that, by setting a `ocspResponderURL` configuration field. If
set, then Pebble will add to the certificate template the given OCSP Responder
URL, to be included in the certificates when they are generated. If not set
(empty string), the template does not contain the OCSP field.

Typical example is:
`"ocspResponderURL": "http://127.0.0.1:4002 pebble"`

Instruction in README are updated with this new config variable.