pebble: expose a separate management interface (#248)
This PR adds a management interface, that listens to a specific address and
port as defined by the field `managementListenAddress` in the Pebble JSON
configuration file, and serves endpoints unrelated to the ACME protocol, but
useful to interface Pebble in a testing environment. This interface is reached
through HTTPS and is secured using the same certificate than the ACME
interface.

Not setting `managementListenAddress` disables the management interface.