Add PEBBLE_CHAIN_LENGTH. (#320)
This commit adds a way to control how long the certificate issuance
chains that Pebble uses are. The default and minimum value of 1 means
that there is one intermediate certificate in the final certificate
chain.

Fixes #318.


The resulting certificate chain (at a value of `2`) looks like this:

    $ cat b.pem | openssl crl2pkcs7 -nocrl -certfile /dev/stdin | openssl pkcs7 -print_certs -noout
    subject=CN = xfoobazx.com
    issuer=CN = Pebble Intermediate CA 2f1dd1
    subject=CN = Pebble Intermediate CA 2f1dd1
    issuer=CN = Pebble Intermediate CA 5a564f
    subject=CN = Pebble Intermediate CA 5a564f
    issuer=CN = Pebble Root CA 0335fe

The change is a little bit awkward because Pebble has been, in parts, written with an
assumption about there being one intermediate.

The additional intermediates are not exposed via the management interface. It might
be possible to make a non-breaking change to `:15000/intermediate-keys/`, but I
haven't been able to think of a use-case for exposing them.