Bump the patch level of golangci-lint
I'm not able to identify the root cause, but it appears it may be a binary
incompatibility somewhere ("export data is newer version - update tool").

There is a lot of releases of golangci-lint since v1.24.6, however this is
the minimum bump required to proceed.  It would probably be a good idea to
update, however the goal of this patch is to pass CI with a minimum amount
of change.

It's not clear why v1.24.4 and v1.24.5 fail to install.