Validate CSR signatures and check signature type (#386)
With the deprecation of SHA1 signatures on CSRs, we should ensure
Pebble rejects them as well. As well, ensure the signature itself is
valid. Unlike Boulder, which currently allows SHA1 to be configurably
accepted, Pebble will always reject SHA1.

There's no tests for the WFE code in this repository, so I didn't add a
unit test ensuring SHA1 is rejected. However, I tested this by running an
ACME client against Pebble that was patched to use SHA1 signatures and
verifying they were successfully rejected, as well as ensuring SHA256
was still accepted.

Fixes #384