Update README.md for ghcr.io docker images (#450)
Pebble is now being published to the Github container registry as part
of the release process.

Update the README to link to it instead of docker hub. Fix the docker
examples to use ghcr.io.

Remove the note about multi-stage builds requiring docker ce 17.05 -
this is now a widely used feature, and the mentioned Docker version is 7
years old, so we shouldn't need a special note here.