Require HS256, HS384, or HS512 for EAB (#459)
During the go-jose/v4 upgrade, I accidentally required the same
signature sets for EAB as for the account keys, which is incorrect. This
allows the correct MAC-based algorithms. It drops the custom algorithm
checks, which are now unreachable as go-jose will enforce the
algorithms.

This also adds a new EAB key to Pebble's test config which explicitly
has base64url characters, from #428 

Fixes #455

---------

Co-authored-by: Folke Gleumes <folke@gleumes.org>