return logical and compliant ARI windows for expiring certs (#484)
While testing ARI functionality with extremely short lived certs, Pebble
returned illogical data - such as the windowEnd appearing in the future
for already expired certs, or after the cert's notAfter.

This PR adjusts `RenewalInfoSimple` to do the following:

* if the cert has expired, just return `RenewalInfoImmediate`
* if the cert will expire before the `windowEnd`, set the `windowEnd` to
the cert's `notAfter`
* cleanup: RFC requires the `windowStart` to be before `windowEnd`, so
remove a second from the `windowStart` if needed.

I also made the `RenewalInfoImmediate` window change from `-60:-30`
minutes to `-60:-59.59` minutes; the sole reason for this change is that
it is visually easier to see the change (and not need to do mental math)
and recognize this is a `RenewalInfoImmediate` from logs.